+++ web/src/main/scala/sc4pac/web/ChannelPage.scala
@@ -45,7 +45,6 @@ object ChannelPage {
   // val sc4pacUrl = "https://github.com/memo33/sc4pac-tools#sc4pac"
   val sc4pacUrl = "https://memo33.github.io/sc4pac/#/"
   val issueUrl = "https://github.com/memo33/sc4pac/issues"
-  val yamlUrl = "https://github.com/memo33/sc4pac/tree/main/src/yaml/"
 
   lazy val backend = sttp.client4.fetch.FetchBackend()
 
@@ -159,7 +158,7 @@ object ChannelPage {
 
     H.div(
       H.div(H.float := "right")(
-        pkg.metadataSource.toSeq.map(yamlSubPath => H.a(H.cls := "btn", H.href := s"$yamlUrl$yamlSubPath")("Edit metadata"))
+        pkg.metadataSourceUrl.toSeq.map(yamlUrl => H.a(H.cls := "btn", H.href := yamlUrl.toString)("Edit metadata"))
         :+ H.a(H.cls := "btn", H.href := issueUrl)("Report a problem")
       ),
       H.h2(H.clear := "right")(module.orgName),